Refactored the functions for fomatting files as openpyxl workbooks and redistributed functions in new modules
