+++ bmfuncts/__init__.py
@@ -10,11 +10,14 @@ from bmfuncts.useful_functs import *
 from bmfuncts.employees_globals import *
 from bmfuncts.pub_globals import *
 from bmfuncts.rename_cols import *
-from bmfuncts.use_pub_attributes import *
+from bmfuncts.use_homonyms import *
+from bmfuncts.use_otps import *
 from bmfuncts.update_employees import *
 from bmfuncts.merge_pub_employees import *
 from bmfuncts.consolidate_pub_list import *
 from bmfuncts.update_impact_factors import *
-from bmfuncts.pub_analysis import *
+from bmfuncts.impact_factors_analysis import *
 from bmfuncts.authors_analysis import *
+from bmfuncts.coupling_analysis import *
+from bmfuncts.keywords_analysis import *
 from bmfuncts.save_final_results import *
