+++ bmfuncts/config_utils.py
@@ -247,6 +247,12 @@ def set_org_params(institute, bibliometer_path):
     that has not been identified as Institute employee.
     - index 10 = the status of droping particular affiliation authors in the file of list of \
     publications with one row per author that has not been identified as Institute employee.
+    - index 11 = the level at which the OTPs are predefined before final set by the user.
+    - index 12 = the name of the database file of OTPs per departement, service and labs.
+    - index 13 = the name of the sheet to be red in the database file of OTPs.
+    - index 14 = the lines number of the header in the database file of OTPs.
+    - index 15 = the columns name to be red in the database file of OTPs.
+    - index 16 = the list of departments that have not lab-OTPs available.
 
     Args:
         institute (str): The Intitute name.
@@ -288,10 +294,17 @@ def set_org_params(institute, bibliometer_path):
     orphan_split_status = inst_org_dict["ORPHAN_SPLIT_STATUS"]
     affil_drop_dict = inst_org_dict["AFFIL_DROP_DICT"]
     orphan_drop_dict = dict(zip(inst_col_list, affil_drop_dict.values()))
+    otps_level = inst_org_dict["OTPS_LEVEL"]
+    lab_otps_bdd = inst_org_dict["LAB_OTPS_BDD"]
+    otps_sheet = inst_org_dict["OTPS_SHEET"]
+    otps_header = inst_org_dict["OTPS_HEADER"]
+    otps_cols = inst_org_dict["OTPS_COL"]
+    nolab_depts = inst_org_dict["NO_LAB_DEPTS"]
 
     return_tup = (col_names_dpt, dpt_label_dict, dpt_attributes_dict,
                   institutions_filter_list, inst_col_list,
                   if_db_status, no_if_doctype_keys_list,
                   main_inst_idx, and_inst_status, orphan_split_status,
-                  orphan_drop_dict)
+                  orphan_drop_dict, otps_level, lab_otps_bdd,
+                  otps_sheet, otps_header, otps_cols, nolab_depts)
     return return_tup
