+++ bmgui/consolidate_corpus_page.py
@@ -29,16 +29,16 @@ import bmfuncts.employees_globals as eg
 import bmfuncts.pub_globals as pg
 import bmgui.gui_globals as gg
 from bmfuncts.config_utils import set_org_params
-from bmfuncts.consolidate_pub_list import add_otp
 from bmfuncts.consolidate_pub_list import built_final_pub_list
 from bmfuncts.consolidate_pub_list import concatenate_pub_lists
 from bmfuncts.consolidate_pub_list import solving_homonyms
 from bmfuncts.merge_pub_employees import recursive_year_search
 from bmfuncts.update_employees import set_employees_data
 from bmfuncts.update_employees import update_employees
-from bmfuncts.use_pub_attributes import save_homonyms
-from bmfuncts.use_pub_attributes import set_saved_homonyms
-from bmfuncts.use_pub_attributes import set_saved_otps
+from bmfuncts.use_homonyms import save_homonyms
+from bmfuncts.use_homonyms import set_saved_homonyms
+from bmfuncts.use_otps import add_otp
+from bmfuncts.use_otps import set_saved_otps
 from bmfuncts.useful_functs import check_dedup_parsing_available
 from bmgui.gui_globals import GUI_BUTTONS
 from bmgui.gui_utils import disable_buttons
@@ -437,14 +437,14 @@ def _launch_add_otp_try(institute, org_tup,
     """Launches files creation for adding OTP attribute to publications.
 
     This is done through the `add_otp` function imported from 
-    `bmfuncts.consolidate_pub_list` module after:
+    `bmfuncts.use_otps` module after:
 
     - check of status of homonyms resolution step 
     - saving the resolved homonyms through `save_homonyms` function \
     imported from `bmfuncts.use_pub_attributes` module.
 
     The created files are filled with previously set OTPs through 
-    `set_saved_otps` function imported from `bmfuncts.use_pub_attributes` 
+    `set_saved_otps` function imported from `bmfuncts.use_otps` 
     module. 
 
     Args:
